// reference/Texas-FrontEnd-TexasInformativeWebsite-release/src/types/resources.ts
export interface ContactPageResourcesProps {
  Home: string;
  Contactus: string;
  contact: string;
  "send-us-a-message": string;
  "name-is-required": string;
  "name-must-be-at-most": string;
  "characters-long": string;
  "name-can-only-contain-letters-or-abostrophe": string;
  "phone-number-is-required": string;
  "phone-number-must-be-at-most": string;
  "email-is-required": string;
  "invalid-email-address": string;
  "please-select-location": string;
  "please-select-inquiry-type": string;
  "date-required": string;
  "date-must-current-past-date": string;
  "time-required": string;
  "invalid-time": string;
  "subject-must-be-between": string;
  and: string;
  "message-is-required": string;
  "captcha-is-required": string;
  "successfully-submitted": string;
  name: string;
  "phone-number": string;
  Email: string;
  state: string;
  "nearest-location": string;
  TypeofInquiry: string;
  "preferred-date": string;
  "preferred-time": string;
  placeholder_Subject: string;
  placeholder_Message: string;
  Submit: string;
  "file-size-exceeds-the-limit-of-5mb": string;
  "invalid-file-type-only-are-allowed": string;
  "you-can-only-upload-maximum": string;
  files: string;
  "total-files-size-exceeds-the-limit": string;
  "upload-files": string;
  "drag-and-drop": string;
  "max-files-total": string;
  "time-of-incident": string;
}